videofetcher/requirements.txt
requests
telegram
python-telegram-bot
yt-dlp
speedtest_cli
